summary_news/requirements.txt
absl-py==0.10.0
APScheduler==3.6.3
astor==0.8.1
astroid==2.4.2
attrs==20.2.0
Automat==20.2.0
beautifulsoup4==4.9.3
blis==0.4.1
boto3==1.15.17
botocore==1.18.17
bs4==0.0.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cloudpickle==0.6.1
colorama==0.4.4
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
ConfigArgParse==0.13.0
constantly==15.1.0
cryptography==3.1.1
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
decorator==4.4.2
docopt==0.6.2
fakeredis==0.10.3
fbmessenger==5.6.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Simple==0.0.3
future==0.17.1
gast==0.4.0
gensim==3.8.3
gevent==1.5.0
greenlet==0.4.17
grpcio==1.32.0
h5py==2.10.0
humanfriendly==8.2
hyperlink==20.0.1
idna==2.10
importlib-metadata==2.0.0
incremental==17.5.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonpickle==1.4.1
jsonschema==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
klein==17.10.0
lazy-object-proxy==1.4.3
Markdown==3.3.1
MarkupSafe==1.1.1
matplotlib==2.2.5
mattermostwrapper==2.2
mccabe==0.6.1
mock==4.0.2
murmurhash==1.0.2
networkx==2.5
nltk==3.5
numpy==1.19.2
packaging==18.0
pika==0.12.0
plac==1.1.3
preshed==3.0.2
prompt-toolkit==3.0.8
protobuf==3.13.0
pycparser==2.20
pydot==1.4.1
PyHamcrest==2.0.2
PyJWT==1.7.1
pykwalify==1.7.0
pylint==2.6.0
pymongo==3.11.0
pyparsing==2.4.7
pyreadline==2.1
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==3.1.2
python-telegram-bot==11.1.0
pytz==2018.9
pyvi==0.1
PyYAML==5.3.1
questionary==1.7.0
rasa-core==0.14.5
rasa-core-sdk==0.14.0
rasa-nlu==0.15.1
redis==2.10.6
regex==2020.10.15
requests==2.24.0
requests-toolbelt==0.9.1
rocketchat-API==0.6.36
ruamel.yaml==0.15.100
s3transfer==0.3.3
scikit-learn==0.20.4
scipy==1.5.2
simplejson==3.17.2
six==1.15.0
sklearn-crfsuite==0.3.6
slackclient==1.3.2
smart-open==3.0.0
soupsieve==2.0.1
spacy==2.3.2
srsly==1.0.2
tabulate==0.8.7
tensorboard==1.13.1
tensorflow==1.13.2
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminaltables==3.1.0
thinc==7.4.1
toml==0.10.1
tqdm==4.50.2
twilio==6.46.0
Twisted==20.3.0
typed-ast==1.4.1
typing==3.7.4.3
tzlocal==2.1
urllib3==1.25.10
wasabi==0.8.0
wcwidth==0.2.5
webexteamssdk==1.6
websocket-client==0.54.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.1
zope.interface==5.1.2
